🎨 Merge modules into main.js
Merged all modules back into main.js because I rain into a bug with the functions being in seperate files that I couldn't fix
